refac: changelog api
